requirements.txt
biopython==1.82
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.0
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
polars==0.20.3
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.12.25
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
StringDist==1.0.9
threadpoolctl==3.2.0
tzdata==2023.4
XlsxWriter==3.1.9
